Issue #93 - Have explicit naming for integers and remove brigand::false_ and brigand::true_

+++ test/repeat_test.cpp
@@ -2,14 +2,14 @@
 #include <brigand/functions/arithmetic/next.hpp>
 #include <brigand/types/integer.hpp>
 
-static_assert(brigand::repeat<brigand::next, brigand::int_<0>, brigand::int_<0>>::value == 0, "invalid result");
-static_assert(brigand::repeat<brigand::next, brigand::int_<1>, brigand::int_<0>>::value == 1, "invalid result");
-static_assert(brigand::repeat<brigand::next, brigand::int_<2>, brigand::int_<0>>::value == 2, "invalid result");
-static_assert(brigand::repeat<brigand::next, brigand::int_<3>, brigand::int_<0>>::value == 3, "invalid result");
-static_assert(brigand::repeat<brigand::next, brigand::int_<4>, brigand::int_<0>>::value == 4, "invalid result");
-static_assert(brigand::repeat<brigand::next, brigand::int_<5>, brigand::int_<0>>::value == 5, "invalid result");
-static_assert(brigand::repeat<brigand::next, brigand::int_<6>, brigand::int_<0>>::value == 6, "invalid result");
-static_assert(brigand::repeat<brigand::next, brigand::int_<7>, brigand::int_<0>>::value == 7, "invalid result");
-static_assert(brigand::repeat<brigand::next, brigand::int_<8>, brigand::int_<0>>::value == 8, "invalid result");
-static_assert(brigand::repeat<brigand::next, brigand::int_<9>, brigand::int_<0>>::value == 9, "invalid result");
+static_assert(brigand::repeat<brigand::next, brigand::uint32_t<0>, brigand::uint32_t<0>>::value == 0, "invalid result");
+static_assert(brigand::repeat<brigand::next, brigand::uint32_t<1>, brigand::uint32_t<0>>::value == 1, "invalid result");
+static_assert(brigand::repeat<brigand::next, brigand::uint32_t<2>, brigand::uint32_t<0>>::value == 2, "invalid result");
+static_assert(brigand::repeat<brigand::next, brigand::uint32_t<3>, brigand::uint32_t<0>>::value == 3, "invalid result");
+static_assert(brigand::repeat<brigand::next, brigand::uint32_t<4>, brigand::uint32_t<0>>::value == 4, "invalid result");
+static_assert(brigand::repeat<brigand::next, brigand::uint32_t<5>, brigand::uint32_t<0>>::value == 5, "invalid result");
+static_assert(brigand::repeat<brigand::next, brigand::uint32_t<6>, brigand::uint32_t<0>>::value == 6, "invalid result");
+static_assert(brigand::repeat<brigand::next, brigand::uint32_t<7>, brigand::uint32_t<0>>::value == 7, "invalid result");
+static_assert(brigand::repeat<brigand::next, brigand::uint32_t<8>, brigand::uint32_t<0>>::value == 8, "invalid result");
+static_assert(brigand::repeat<brigand::next, brigand::uint32_t<9>, brigand::uint32_t<0>>::value == 9, "invalid result");
